Move SPIRC into librespot so it can be re-used by other applications
